refactor: Player.x & Player.y -> Player.pos

+++ client/core/src/main/kotlin/me/grian/network/packets/s2c/S2CLoginAcceptedPacket.kt
@@ -12,6 +12,6 @@ class S2CLoginAcceptedPacket : S2CPacket {
         val initialX = readChannel.readInt()
         val initialY = readChannel.readInt()
 
-        Main.player.move(initialX.toFloat(), initialY.toFloat())
+        Main.player.move(initialX, initialY)
     }
 }
